Fix Languages Task Output Path
Despite the output path being set in the config, it was never used in
the task (my error) and so the output path was always going to be
`/languages`. This commit fixes the output path to that defined in
config, and places it under `/develop/languages/` which is intentionally
ignored in our `.gitignore`.

Fixes #4